* commit '5c0b166e90e8c90c5d661a78b8bd4c1e9b531e5f':
  remove org unit
  killing consoles
  styling
  add datatype selector tree to auth source panel
  - datatype tree with search - fix on remote (init)
  moving components to common
  - add to form - styling buttons
  better error msg
  - Use modes to control view - make edit and create same panel - move doUpdate, doSave etc menthods into angular
  Rework Auth Source Section to work with Svelte

+++ waltz-service/src/main/java/com/khartec/waltz/service/authoritative_source/AuthoritativeSourceService.java
@@ -18,6 +18,7 @@
 
 package com.khartec.waltz.service.authoritative_source;
 
+import com.khartec.waltz.common.exception.NotFoundException;
 import com.khartec.waltz.data.GenericSelector;
 import com.khartec.waltz.data.GenericSelectorFactory;
 import com.khartec.waltz.data.application.ApplicationDao;
@@ -25,8 +26,8 @@ import com.khartec.waltz.data.application.ApplicationIdSelectorFactory;
 import com.khartec.waltz.data.authoritative_source.AuthoritativeSourceDao;
 import com.khartec.waltz.data.data_type.DataTypeDao;
 import com.khartec.waltz.data.data_type.DataTypeIdSelectorFactory;
-import com.khartec.waltz.data.orgunit.OrganisationalUnitDao;
 import com.khartec.waltz.data.datatype_decorator.LogicalFlowDecoratorDao;
+import com.khartec.waltz.data.orgunit.OrganisationalUnitDao;
 import com.khartec.waltz.model.*;
 import com.khartec.waltz.model.application.Application;
 import com.khartec.waltz.model.authoritativesource.*;
@@ -143,8 +144,12 @@ public class AuthoritativeSourceService {
 
     public int remove(long id, String username) {
 
-        logRemoval(id, username);
         AuthoritativeSource authSourceToDelete = getById(id);
+
+        if(authSourceToDelete == null){
+            throw new NotFoundException("ASRM-NF", "Authoritative source not found");
+        }
+        logRemoval(id, username);
         int deletedCount = authoritativeSourceDao.remove(id);
         ratingCalculator.update(authSourceToDelete.dataType(), authSourceToDelete.parentReference());
         return deletedCount;
